requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
rdflib==7.0.0
authlib==1.3.0
requests==2.31.0
python-dotenv==1.0.0
Flask-Babel==4.0.0
